requirements.txt
watchdog
plyer
pyyaml 